+++ src/main/java/de/jjohannes/gradle/moduledependencies/GAV.java
@@ -1,7 +1,0 @@
-package de.jjohannes.gradle.moduledependencies;
-
-interface GAV {
-    String GROUP = "group";
-    String ARTIFACT = "name";
-    String VERSION = "version";
-}